Extract OtherOptions to component.
