+++ editor/src/App.svelte
@@ -21,6 +21,7 @@
   import MidiContext from "./components/MidiContext.svelte";
   import MidiEnabled from "./components/MidiEnabled.svelte";
   import MidiSelector from "./components/MidiSelector.svelte";
+  import OtherOptions from "./components/OtherOptions.svelte";
   import Subhead from "./components/Subhead.svelte";
   import { Tabs, TabList, TabPanel, Tab } from './components/tabs';
 
@@ -47,6 +48,7 @@
         OxionMidi.requestConfig($selectedMidiOutput);
         break;
       case "importConfig":
+        // TODO: can you extract this to a file?
         let fileInputNode = document.createElement('input');
         fileInputNode.setAttribute("type", "file");
         fileInputNode.setAttribute("id", "uploadedConfig");
@@ -84,6 +86,7 @@
         fileInputNode.remove();
         break;
       case "exportConfig":
+        // TODO extract to file
         let dataStr = "data:text/json;charset=utf-8," + encodeURIComponent($configuration.toJsonString());
 
         let downloadAnchorNode = document.createElement('a');
@@ -117,9 +120,6 @@
     editMode = !editMode;
   }
 
-  function touchControl() {
-    editConfiguration.update(old => $editConfiguration);
-  }
 </script>
 
 <style>
@@ -146,11 +146,6 @@
 
   p.device {
     text-align: right;
-  }
-
-  p.note {
-    width: 400px;
-    line-height: 1.2;
   }
 
   main {
@@ -213,50 +208,7 @@
             </TabPanel>
 
             <TabPanel>
-              {#if $configuration.device().capabilities.led}
-              <CheckOption bind:checked={$editConfiguration.ledOn}>
-                LED permanently on when powered
-              </CheckOption>
-
-              <CheckOption bind:checked={$editConfiguration.ledFlash}>
-                LED flash on MIDI activity
-              </CheckOption>
-              {/if}
-
-              <CheckOption bind:checked={$editConfiguration.controllerFlip}>
-                Rotate controller 180º
-              </CheckOption>
-
-              {#if $configuration.device().capabilities.i2c}
-              <h3>I2C Master/Follower</h3>
-              <select bind:value={$editConfiguration.i2cMaster} on:change={touchControl}>
-                <option value={false}>Follower</option>
-                <option value={true}>Master</option>
-              </select>
-
-              <p>This will not take effect until you restart (disconnect/reconnect) your 16n.</p>
-
-              <h3>Fader Minimum/Maximum scalars</h3>
-              <div>
-                <label>Fader Minimum read value</label>
-                <input
-                  type="number"
-                  bind:value={$editConfiguration.fadermin}
-                  on:change={touchControl}
-                  min="0"
-                  max="{(1 << 13) - 1}" />
-              </div>
-              <div>
-                <label>Fader Maximum read value</label>
-                <input
-                  type="number"
-                  bind:value={$editConfiguration.fadermax}
-                  on:change={touchControl}
-                  min="0"
-                  max="{(1 << 13) - 1}" />
-              </div>
-              <p class='note'>You probably shouldn't touch this, unless you're having issues with your fader scaling. "Raw" analog read values are from 0 to 8192. The defaults are based on experience.</p>
-              {/if}
+              <OtherOptions />
             </TabPanel>
           </Tabs>
